+++ kedro_mlflow/config/kedro_mlflow_config.py
@@ -4,7 +4,6 @@ from typing import Dict, List, Optional
 from urllib.parse import urlparse
 
 import mlflow
-from kedro.config import MissingConfigException
 from kedro.framework.context import KedroContext
 from kedro.framework.startup import _is_project
 from mlflow.entities import Experiment
@@ -211,15 +210,3 @@ class KedroMlflowConfig(BaseModel):
 
 class KedroMlflowConfigError(Exception):
     """Error occurred when loading the configuration"""
-
-
-def get_mlflow_config(context: KedroContext):
-    try:
-        conf_mlflow_yml = context.config_loader.get("mlflow*", "mlflow*/**")
-    except MissingConfigException:
-        raise KedroMlflowConfigError(
-            "No 'mlflow.yml' config file found in environment. Use ``kedro mlflow init`` command in CLI to create a default config file."
-        )
-    conf_mlflow_yml["project_path"] = context.project_path
-    mlflow_config = KedroMlflowConfig.parse_obj(conf_mlflow_yml)
-    return mlflow_config